# src/main/resources/openapi/noteSharingApp.yaml
openapi: 3.0.3
info:
  title: Note Sharing Application - API Documentation
  description: |-
    This is the documentation for Note Sharing Application API. Made for course Information System Modelling 
  termsOfService: http://swagger.io/terms/
  contact:
    email: 263930@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://note-sharing-app.pl/api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        roles:
          type: array
          items:
            type: string
        user_status:
          type: string
          enum: [active, blocked, banned]
...
paths:
...
  /users/{id}:
    get:
      summary: Get user details
      description: Retrieve information about a specific user. Only available to logged-in users.
      tags:
        - users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Unauthorized - User not logged in
        "403":
          description: Forbidden - Insufficient permissions
        "404":
          description: User not found
        "500":
          description: Internal server error

    delete:
      summary: Delete a user
      description: Delete a specific user. Only available to administrators.
      tags:
        - users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the user to delete
      responses:
        "200":
          description: User successfully deleted
        "401":
          description: Unauthorized - User not logged in
        "403":
          description: Forbidden - Admin access required
        "404":
          description: User not found
        "500":
          description: Internal server error
...